+++ R/utils.R
@@ -1,111 +1,110 @@
-#' \[experimental\] Find analogical relationships between words
+#' Convert formula to named character vector
 #' 
-#' @param x a `textmodel_wordvector` object.
+#' Convert a formula to a named character vector in analogy tasks.
 #' @param formula a [formula] object that defines the relationship between words 
 #'   using `+` or `-` operators.
-#' @param n the number of words in the resulting object.
-#' @param exclude if `TRUE`, words in `formula` are excluded from the result.
-#' @param type specify the type of vectors to be used. "word" is word vectors  
-#'   while "simil" is similarity vectors.
-#' @return a `data.frame` with the words sorted and their cosine similarity sorted 
-#'   in descending order.
-#' @importFrom utils head tail
 #' @export
-#' @references 
-#'   Mikolov, T., Sutskever, I., Chen, K., Corrado, G., & Dean, J. (2013). 
-#'   Distributed Representations of Words and Phrases and their Compositionality. 
-#'   http://arxiv.org/abs/1310.4546.
+#' @seealso [similarity()]
+#' @importFrom utils head tail
+#' @return a named character vector to be passed to [similarity()].
 #' @examples
-#' \donttest{
-#' # from Mikolov et al. (2023)
-#' analogy(wdv, ~ berlin - germany + france)
-#' analogy(wdv, ~ quick - quickly + slowly)
-#' }
-analogy <- function(x, formula, n = 10, exclude = TRUE, type = c("word", "simil")) {
+#' analogy(~ berlin - germany + france)
+#' analogy(~ quick - quickly + slowly)
+analogy <- function(formula) {
     
-    if (!identical(class(x), "textmodel_wordvector"))
-        stop("x must be a textmodel_wordvector object")
-        
-    n <- check_integer(n)
-    exclude <- check_logical(exclude)
-    type <- match.arg(type)
-    emb <- as.matrix(x)
     if (!identical(class(formula), "formula"))
-        stop("The object for 'formula' should be a formula")
+        stop("formula must be a formula object")
     
     f <- tail(as.character(formula), 1)
     match <- stringi::stri_match_all_regex(f, "([+-])?\\s*(\\w+)")[[1]]
     match[,2] <- stringi::stri_trim(match[,2])
     match[,2][is.na(match[,2])] <- "+"
-    weight <- numeric()
+    res <- numeric()
     for (i in seq_len(nrow(match))) {
         m <- match[i,]
-        if (!m[3] %in% rownames(emb)) {
-            warning('"', m[3],  '" is not found')
-            next
-        }
         if (m[2] == "-") {
-            weight <- c(weight, structure(-1.0, names = m[3]))
+            res <- c(res, structure(-1.0, names = m[3]))
         } else if (m[2] == "+") {
-            weight <- c(weight, structure(1.0, names = m[3]))
+            res <- c(res, structure(1.0, names = m[3]))
         }
     }
-    
-    j <- match(names(weight), rownames(emb))
-    if (length(j) == 0) {
-        res <- data.frame(word = character(), similarity = numeric())
-    } else {
-        v <- emb[j,, drop = FALSE]
-        if (exclude)
-            emb <- emb[j * -1,, drop = FALSE]
-        if (type == "word") {
-            s <- Matrix::rowMeans(proxyC::simil(emb, t(t(v) %*% weight), use_nan = TRUE))
-        } else {
-            s <- Matrix::rowMeans(proxyC::simil(emb, v, use_nan = TRUE) %*% weight)
-        }
-        s <- head(sort(s, decreasing = TRUE), n)
-        res <- data.frame(word = names(s), similarity = s)
-    }
-    rownames(res) <- NULL
-    attr(res, "formula") <- formula
-    attr(res, "weight") <- weight
     return(res)
 }
 
-
 #' Compute similarity between word vectors
 #' 
+#' Compute cosine similarity between word vectors for selected words.
 #' @param x a `textmodel_wordvector` object.
 #' @param words words for which similarity is computed.
 #' @param mode specify the type of resulting object.
-#' @return a `matrix` of cosine similarity scores when `mode = "simil"` or of 
-#'   words sorted by the similarity scores when `mode = "word`.
+#' @return a `matrix` of cosine similarity scores when `mode = "values"` or of 
+#'   words sorted in descending order by the similarity scores when `mode = "words"`.
+#'   When `words` is a named numeric vector, word vectors are weighted and summed 
+#'   before computing similarity scores.
 #' @export
-similarity <- function(x, words, mode = c("simil", "word")) {
+#' @seealso [analogy()]
+similarity <- function(x, words, mode = c("words", "values")) {
     
     if (!identical(class(x), "textmodel_wordvector"))
         stop("x must be a textmodel_wordvector object")
-    
-    words <- check_character(words, max_len = Inf)
+
     mode <- match.arg(mode)
-    emb <- as.matrix(x)
+    emb1 <- as.matrix(x)
     
-    b <- words %in% rownames(emb)
+    if (is.character(words)) {
+        words <- structure(rep(1.0, length(words)), names = words)
+        weighted <- FALSE
+    } else if (is.numeric(words)) {
+        if (is.null(names(words)))
+            stop("words must be named")
+        weighted <- TRUE 
+    } else {
+        stop("words must be a character or named numeric vector")
+    }
+    b <- names(words) %in% rownames(emb1)
     if (sum(!b) == 1) {
-        warning(paste0('"', words[!b], '"',  collapse = ", "),  ' is not found')
+        warning(paste0('"', names(words[!b]), '"',  collapse = ", "),  ' is not found')
     } else if (sum(!b) > 1) {
-        warning(paste0('"', words[!b], '"',  collapse = ", "),  ' are not found')
+        warning(paste0('"', names(words[!b]), '"',  collapse = ", "),  ' are not found')
     }
     words <- words[b]
-    res <- as.matrix(proxyC::simil(emb, emb[words,, drop = FALSE], use_nan = TRUE))
+    if (weighted) {
+        emb2 <- rbind(colSums(emb1[names(words),, drop = FALSE] * words))
+    } else {
+        emb2 <- emb1[names(words),, drop = FALSE]
+    }
+    res <- as.matrix(proxyC::simil(emb1, emb2, use_nan = TRUE))
     if (ncol(res) == 0) {
         res <- matrix(nrow = 0, ncol = 0)
     } else {
-        if (mode == "word") {
+        if (mode == "words") {
             res <- apply(res, 2, function(v) {
                 names(sort(v, decreasing = TRUE))
             })
         }
+    }
+    return(res)
+}
+
+#' \[experimental\] Extract word vector weights
+#' 
+#' @param x a `textmodel_wordvector` object.
+#' @param mode specify the type of resulting object.
+#' @return a `matrix` of word vector weights when `mode = "value"` or of 
+#'   words sorted in descending order by the weights when `mode = "word"`.
+#' @export
+weights <- function(x, mode = c("words", "values")) {
+    
+    if (!identical(class(x), "textmodel_wordvector"))
+        stop("x must be a textmodel_wordvector object")
+    
+    mode <- match.arg(mode)
+    if (mode == "values") {
+        res <- x$weights
+    } else {
+        res <- apply(x$weights, 2, function(x) {
+            names(sort(x, decreasing = TRUE))
+        })
     }
     return(res)
 }